separate root index of shared login UX
