+++ packages/core/src/login/webview/vue/amazonq/backend_amazonq.ts
@@ -3,13 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import * as vscode from 'vscode'
-import { SsoConnection } from '../../../auth/connection'
-import { AuthUtil } from '../../../codewhisperer/util/authUtil'
-import { AuthError, CommonAuthWebview } from './backend'
-import { awsIdSignIn } from '../../../codewhisperer/util/showSsoPrompt'
-import { connectToEnterpriseSso } from '../../../codewhisperer/util/getStartUrl'
+import { SsoConnection } from '../../../../auth/connection'
+import { AuthUtil } from '../../../../codewhisperer/util/authUtil'
+import { AuthError, CommonAuthWebview } from '../backend'
+import { awsIdSignIn } from '../../../../codewhisperer/util/showSsoPrompt'
+import { connectToEnterpriseSso } from '../../../../codewhisperer/util/getStartUrl'
 
 export class AmazonQLoginWebview extends CommonAuthWebview {
+    public override source: string = 'src/login/webview/vue/amazonq/index.js'
+
     fetchConnection(): SsoConnection | undefined {
         if (AuthUtil.instance.isConnected() && AuthUtil.instance.conn?.type === 'sso') {
             return AuthUtil.instance.conn
